+++ src/main/java/org/opensearch/knn/index/util/KNNLibrary.java
@@ -69,6 +69,16 @@ public interface KNNLibrary {
     float score(float rawScore, SpaceType spaceType);
 
     /**
+     * Translate the distance radius input from end user to the engine's threshold.
+     *
+     * @param distance distance radius input from end user
+     * @param spaceType spaceType used to compute the radius
+     *
+     * @return transformed distance for the library
+     */
+    Float distanceToRadialThreshold(Float distance, SpaceType spaceType);
+
+    /**
      * Validate the knnMethodContext for the given library. A ValidationException should be thrown if the method is
      * deemed invalid.
      *